[Feature] ARMA Anomaly Detection
The system now detects anomalies based on stddev of the ARMA movement
data. This seems to be a compliment to normal StdDev as they pick up a
lot of common data however they also seem to pickup a different data
set, which seems to mean that both are useful in different ways.